+++ wgan_model_and_callback.py
@@ -1,8 +1,7 @@
 """
-model definitions and callbacks
+wgan-gp model definition and callback
 
-Implementation modified from following this tutorial:
-https://keras.io/examples/generative/wgan_gp/
+wgan-gp combines the critic and generator into one model for training.
 """
 
 
@@ -10,11 +9,8 @@ import os
 import time
 from datetime import datetime
 import shutil
-from pathlib import Path
 import pandas as pd
-import numpy as np
 import keras
-from keras import layers
 import tensorflow as tf
 import matplotlib
 import matplotlib.pyplot as plt
@@ -35,441 +31,6 @@ warnings.filterwarnings(
 
 # Clear all previously registered custom objects. Necessary addition for loading models with custom objects.
 keras.saving.get_custom_objects().clear()
-
-@keras.saving.register_keras_serializable(package="critic_model", name="critic_model")
-def get_critic_model(img_shape: tuple, num_classes: int, model_training_output_dir: Path):
-    """
-    initialize the critic using the Keras Functional API. The critic model takes as input an image tensor and a label tensor and outputs a single
-    value. The label tensor is converted to a one-hot tensor, passed through a fully connected layer to reshape it to the same shape as the image
-    and then concatenated with the image tensor. This combined tensor is passed through the convolutional layers and lastly through a dense layer
-    without an activation function to output the critic's score for the input image-label pair.
-    
-    I removed dropout from the critic model, but could add it back if the model is overfitting.
-    
-    This function also:
-        1. saves a copy of the model summary to a text file in the model_training_output directory
-        2. visualizes the model architecture with keras.utils.plot_model
-        3. saves a copy of the model for visualization with netron.app
-    
-    Parameters:
-        img_shape: tuple, the shape of the input image tensor (height, width, channels)
-        num_classes: int, the number of classes in the dataset
-        model_training_output_dir: Path, the directory to save outputs from the model training
-    Returns:
-        critic_model: keras.Model, the compiled critic model
-    """
-    # ############################################################### Input Preparation ###############################################################
-    # # Image input
-    # img_input = layers.Input(shape=img_shape)
-    
-    # # class input
-    # class_input = layers.Input(shape=(1,), dtype=tf.int32)  # Assuming labels are integers
-    # class_input_one_hot = layers.Embedding(num_classes, num_classes)(class_input)  # Convert to one-hot
-    # class_input_one_hot = layers.Flatten()(class_input_one_hot)  # Flatten embedding output
-    
-    # # Expand labels to match the image dimensions
-    # # Fully connected layer to reshape one-hot labels to a tensor the same shape as a flattened image
-    # label_tensor = layers.Dense(np.prod(img_shape))(class_input_one_hot)
-    # # Reshape to match image shape (height, width, channels)
-    # label_tensor = layers.Reshape(img_shape)(label_tensor)
-    
-    # # Concatenate image and label tensors
-    # combined_input = layers.Concatenate()([img_input, label_tensor])
-    
-    # ############################################################## Critic Architecture ##############################################################
-    # ########## Convolutional layer 1
-    # x = layers.Conv2D(filters=64, kernel_size=(5, 5), strides=(2, 2), padding="same", use_bias=True)(combined_input)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Convolutional layer 2
-    # x = layers.Conv2D(filters=128, kernel_size=(5, 5), strides=(2, 2), padding="same", use_bias=True)(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Convolutional Layer 3
-    # x = layers.Conv2D(filters=256, kernel_size=(5, 5), strides=(2, 2), padding="same", use_bias=True)(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Flatten the output and add a dense layer with no activation for a single critic score
-    # x = layers.Flatten()(x)
-    # x = layers.Dense(1, use_bias=True)(x)
-    
-    # ########## Define the model with both inputs
-    # critic_model = keras.models.Model([img_input, class_input], x, name="critic")
-    
-    # -----------------------------
-    #   Inputs
-    # -----------------------------
-    img_input = layers.Input(shape=img_shape, name="image_input")
-    class_input = layers.Input(shape=(1,), dtype=tf.int32, name="class_input")
-
-    # Convert label to an embedding vector of size num_classes (or whatever dimension you prefer)
-    label_embed = layers.Embedding(num_classes, num_classes)(class_input)  # shape: (batch,1,num_classes)
-    label_embed = layers.Flatten()(label_embed)                            # shape: (batch, num_classes)
-
-    # -----------------------------
-    #   Conv Block #1
-    #   28×28 → 14×14, 64 filters
-    # -----------------------------
-    x = layers.Conv2D(
-        filters=64, kernel_size=(5,5),
-        strides=(2,2), padding="same", use_bias=True
-    )(img_input)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Residual Block #1 (14×14)
-    # -----------------------------
-    shortcut = x
-    x = layers.Conv2D(
-        filters=64, kernel_size=(3,3), 
-        strides=(1,1), padding="same", use_bias=True
-    )(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    x = layers.Conv2D(
-        filters=64, kernel_size=(3,3), 
-        strides=(1,1), padding="same", use_bias=True
-    )(x)
-    x = layers.Add()([shortcut, x])
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Label Fusion in the Middle
-    #   (Both x and label map have shape involving 14×14)
-    # -----------------------------
-    # Convert label_embed to a 14×14×1 map
-    label_map = layers.Dense(14 * 14, use_bias=False)(label_embed)
-    label_map = layers.Reshape((14, 14, 1))(label_map)
-    # Concatenate along the channels: x is (14,14,64), label_map is (14,14,1)
-    x = layers.Concatenate()([x, label_map])  # Now shape is (14,14,64+1) = (14,14,65)
-
-    # -----------------------------
-    #   Conv Block #2
-    #   14×14 → 7×7, 128 filters
-    # -----------------------------
-    x = layers.Conv2D(
-        filters=128, kernel_size=(5,5),
-        strides=(2,2), padding="same", use_bias=True
-    )(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Residual Block #2 (7×7)
-    # -----------------------------
-    shortcut = x
-    x = layers.Conv2D(
-        filters=128, kernel_size=(3,3),
-        strides=(1,1), padding="same", use_bias=True
-    )(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    x = layers.Conv2D(
-        filters=128, kernel_size=(3,3),
-        strides=(1,1), padding="same", use_bias=True
-    )(x)
-    x = layers.Add()([shortcut, x])
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Conv Block #3
-    #   7×7 → 4×4 (approx), 256 filters
-    # -----------------------------
-    x = layers.Conv2D(
-        filters=256, kernel_size=(5,5),
-        strides=(2,2), padding="same", use_bias=True
-    )(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Flatten + Final Dense → Score
-    # -----------------------------
-    x = layers.Flatten()(x)
-    x = layers.Dense(1, use_bias=True)(x)
-
-    # Build the model
-    critic_model = keras.models.Model([img_input, class_input], x, name="critic_with_mid_label_fusion")
-    return critic_model
-    
-    ######################################################## Model Summary and Visualization ########################################################
-    # get the filepath to the model_architecture_and_summary directory
-    model_architecture_and_summary_dir = model_training_output_dir.joinpath("model_architecture_and_summary")
-    # create the directory if it doesn't exist
-    os.makedirs(model_architecture_and_summary_dir, exist_ok=True)
-    
-    # save a copy of the model summary to a text file
-    model_summary_path = model_architecture_and_summary_dir.joinpath("critic_model_summary.txt")
-    with open(model_summary_path, "w") as file:
-        critic_model.summary(line_length=150, print_fn=lambda x: file.write(x + "\n"))
-    
-    # visualize the model architecture with keras.utils.plot_model
-    model_visualization_path = model_architecture_and_summary_dir.joinpath("critic_model_architecture.png")
-    keras.utils.plot_model(critic_model, to_file=model_visualization_path, show_shapes=True, show_layer_names=True)
-    
-    # save a copy of the model for visualization with netron.app
-    model_netron_path = model_architecture_and_summary_dir.joinpath("critic_model_for_netron_app.keras")
-    critic_model.save(model_netron_path)
-    
-    print(f"\nDone initializing the critic model.")
-    print(f"Model summary saved to: {model_summary_path}")
-    print(f"Visualization of architecture saved to: {model_visualization_path}")
-    print(f"Model saved for visualization with netron.app to: {model_netron_path}\n")
-    return critic_model
-
-
-@keras.saving.register_keras_serializable(package="generator_model", name="generator_model")
-def get_generator_model(noise_dim_shape: int, num_classes: int, model_training_output_dir: Path):
-    """
-    Builds and returns a generator model for a Wasserstein GAN with Gradient Penalty (WGAN-GP).
-    Parameters:
-        noise_dim_shape (int): Dimension of the noise vector input.
-        num_classes (int): Number of classes for the class conditional input.
-        model_training_output_dir (Path): Directory path for saving model training outputs.
-    Returns:
-        keras.models.Model: A Keras Model representing the generator.
-    """
-    ############################################################### Input Preparation ###############################################################
-    # model inputs will look like this: [noise_input, class_input]
-    # Noise input
-    # noise_input = layers.Input(shape=(noise_dim_shape,))
-    
-    # # class input (conditional input)
-    # class_input = layers.Input(shape=(1,), dtype=tf.int32)  # Assuming class_inputs are integers
-    # class_input_one_hot = layers.Embedding(num_classes, num_classes)(class_input)  # Convert to one-hot
-    # class_input_one_hot = layers.Flatten()(class_input_one_hot)  # Flatten embedding output
-    # # REVIEW: consider applying a Dense layer with tanh activation to scale to [-1, 1]
-    # # class_input_one_hot = layers.Dense(num_classes, activation="tanh")(class_input_one_hot)
-    
-    # # Concatenate noise input and class input
-    # x = layers.Concatenate()([noise_input, class_input_one_hot])
-    
-    # ############################################################# Generator Architecture ############################################################
-    # ########## Initial Dense layer
-    # x = layers.Dense(4 * 4 * 256, use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    # x = layers.Reshape((4, 4, 256))(x)
-    
-    # ########## Upsample Block 1
-    # x = layers.UpSampling2D((2, 2))(x)
-    # x = layers.Conv2D(256, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Upsample Block 2
-    # x = layers.UpSampling2D((2, 2))(x)
-    # x = layers.Conv2D(128, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Upsample Block 3
-    # x = layers.UpSampling2D((2, 2))(x)
-    # x = layers.Conv2D(1, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.Activation("tanh")(x)
-    
-    # # At this point, we have an output which has a shape of (32, 32, 1) but we want our output to be (28, 28, 1).
-    # # We will use a Cropping2D layer to make it (28, 28, 1).
-    # x = layers.Cropping2D((2, 2))(x)
-    
-    # ########## Define the model with both inputs
-    # generator_model = keras.models.Model([noise_input, class_input], x, name="generator")
-    
-    # ############################################################ Generator Architecture 2 ###########################################################
-    # ########## Initial Dense layer
-    # x = layers.Dense(7 * 7 * 256, use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    # x = layers.Reshape((7, 7, 256))(x)
-    
-    # ########### Residual block at 14x14 resolution
-    # shortcut = x  # Save the input for the shortcut connection
-    # # First convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    # # Second convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # # Add the shortcut connection
-    # x = layers.Add()([x, shortcut])
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Upsample Block 1 transforms 7x7 -> 14x14
-    # # x = layers.UpSampling2D((2, 2))(x)
-    # # x = layers.Conv2D(256, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # # x = layers.BatchNormalization()(x)
-    # # x = layers.LeakyReLU(0.2)(x)
-    # x = layers.Conv2DTranspose(filters=256, kernel_size=(5,5), strides=(2,2), padding='same', use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########### Residual block at 14x14 resolution
-    # shortcut = x  # Save the input for the shortcut connection
-    # # First convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    # # Second convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # # Add the shortcut connection
-    # x = layers.Add()([x, shortcut])
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Upsample Block 2 transforms 14x14 -> 28x28
-    # # x = layers.UpSampling2D((2, 2))(x)
-    # # x = layers.Conv2D(1, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # # # x = layers.BatchNormalization()(x)
-    # # x = layers.LeakyReLU(0.2)(x)
-    # x = layers.Conv2DTranspose(filters=256, kernel_size=(5,5), strides=(2,2), padding='same', use_bias=False)(x)
-    # x = layers.BatchNormalization()(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########### Residual block at 28x28 resolution
-    # shortcut = x  # Save the input for the shortcut connection
-    # # First convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # x = layers.LeakyReLU(0.2)(x)
-    # # Second convolution
-    # x = layers.Conv2D(filters=256, kernel_size=(3, 3), strides=(1, 1), padding="same")(x)
-    # # Add the shortcut connection
-    # x = layers.Add()([x, shortcut])
-    # x = layers.LeakyReLU(0.2)(x)
-    
-    # ########## Final Layer for Generator
-    # # Use 1 filter to generate a grayscale MNIST image, and apply Tanh activation to scale to [-1, 1].
-    # x = layers.Conv2D(1, (5, 5), strides=(1, 1), padding="same", use_bias=False)(x)
-    # x = layers.Activation("tanh")(x)
-    
-    # ########## Define the model with both inputs
-    # generator_model = keras.models.Model([noise_input, class_input], x, name="generator")
-    
-    ############################################################ Generator Architecture 3 ###########################################################
-    # arguments
-    noise_dim = noise_dim_shape
-    embed_dim = 10
-    # -----------------------------
-    #   Inputs
-    # -----------------------------
-    noise_input = layers.Input(shape=(noise_dim,))
-    class_input = layers.Input(shape=(1,), dtype=tf.int32)
-
-    # Class embedding → Flatten
-    class_embed = layers.Embedding(num_classes, embed_dim)(class_input)
-    class_embed = layers.Flatten()(class_embed)
-
-    # Concatenate noise + embedded class
-    x = layers.Concatenate()([noise_input, class_embed])
-
-    # -----------------------------
-    #   Initial Dense + Reshape
-    # -----------------------------
-    x = layers.Dense(7 * 7 * 256, use_bias=False)(x)  
-    x = layers.BatchNormalization()(x)  
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    x = layers.Reshape((7, 7, 256))(x)
-
-    # -----------------------------
-    #   Residual Block at 7×7
-    # -----------------------------
-    shortcut = x
-    x = layers.Conv2D(256, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-    x = layers.LeakyReLU(0.2)(x)
-
-    x = layers.Conv2D(256, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-
-    x = layers.Add()([x, shortcut])  # residual add
-    x = layers.LeakyReLU(0.2)(x)
-
-    # -----------------------------
-    #   Upsample to 14×14 (256→128 filters)
-    # -----------------------------
-    x = layers.Conv2DTranspose(
-        128, (5,5), strides=(2,2), 
-        padding='same', use_bias=False
-    )(x)
-    x = layers.BatchNormalization()(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Residual Block at 14×14
-    # -----------------------------
-    shortcut = x
-    x = layers.Conv2D(128, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-    x = layers.LeakyReLU(0.2)(x)
-
-    x = layers.Conv2D(128, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-
-    x = layers.Add()([x, shortcut])  # residual add
-    x = layers.LeakyReLU(0.2)(x)
-
-    # -----------------------------
-    #   Upsample to 28×28 (128→64 filters)
-    # -----------------------------
-    x = layers.Conv2DTranspose(
-        64, (5,5), strides=(2,2), 
-        padding='same', use_bias=False
-    )(x)
-    x = layers.BatchNormalization()(x)
-    x = layers.LeakyReLU(alpha=0.2)(x)
-
-    # -----------------------------
-    #   Residual Block at 28×28
-    #   (Optionally remove BN here if you want)
-    # -----------------------------
-    shortcut = x
-    x = layers.Conv2D(64, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-    x = layers.LeakyReLU(0.2)(x)
-
-    x = layers.Conv2D(64, (3,3), padding="same", use_bias=False)(x)
-    x = layers.BatchNormalization()(x)
-
-    x = layers.Add()([x, shortcut])  
-    x = layers.LeakyReLU(0.2)(x)
-
-    # -----------------------------
-    #   Final Conv → 1 channel
-    #   (No BN, Tanh)
-    # -----------------------------
-    x = layers.Conv2D(
-        1, (5,5), padding="same", 
-        use_bias=False  # BN is off
-    )(x)
-    x = layers.Activation("tanh")(x)
-
-    # Define the model
-    generator_model = keras.models.Model(
-        [noise_input, class_input], x, 
-        name="better_generator"
-    )
-    
-    ######################################################## Model Summary and Visualization ########################################################
-    # get the filepath to the model_architecture_and_summary directory
-    model_architecture_and_summary_dir = model_training_output_dir.joinpath("model_architecture_and_summary")
-    # create the directory if it doesn't exist
-    os.makedirs(model_architecture_and_summary_dir, exist_ok=True)
-    
-    # save a copy of the model summary to a text file
-    model_summary_path = model_architecture_and_summary_dir.joinpath("generator_model_summary.txt")
-    with open(model_summary_path, "w") as file:
-        generator_model.summary(line_length=150, print_fn=lambda x: file.write(x + "\n"))
-    
-    # visualize the model architecture with keras.utils.plot_model
-    model_visualization_path = model_architecture_and_summary_dir.joinpath("generator_model_architecture.png")
-    keras.utils.plot_model(generator_model, to_file=model_visualization_path, show_shapes=True, show_layer_names=True)
-    
-    # save a copy of the model for visualization with netron.app
-    model_netron_path = model_architecture_and_summary_dir.joinpath("generator_model_for_netron_app.keras")
-    generator_model.save(model_netron_path)
-    
-    print(f"\nDone initializing the generator model.")
-    print(f"Model summary saved to: {model_summary_path}")
-    print(f"Visualization of architecture saved to: {model_visualization_path}")
-    print(f"Model saved for visualization with netron.app to: {model_netron_path}\n")
-    return generator_model
-
 
 @keras.saving.register_keras_serializable(package="custom_wgan_gp", name="custom_wgan_gp")  # REVIEW: maybe this isn't necessary
 class WGAN_GP(keras.Model):
@@ -804,7 +365,7 @@ class Training_Monitor(tf.keras.callbacks.Callback):
             grid_size: tuple, the size of the grid for the generated images
             samples_per_epoch: int, the number of samples in the training dataset
         """
-        self. model_training_output_dir = model_training_output_dir
+        self.model_training_output_dir = model_training_output_dir
         self.model_checkpoints_dir = model_checkpoints_dir
         self.this_epoch_checkpoint_dir = None  # the path to the current model checkpoint (will be set/updated in on_epoch_end)
         
@@ -930,12 +491,13 @@ class Training_Monitor(tf.keras.callbacks.Callback):
         return
     
     def on_train_end(self, logs=None):
-        epoch = len(self.metrics_dataframe)
-        # skip generating a gif if training ends on a multiple of the frequency to create a gif because it was already generated in the 
-        # generate_validation_samples method at the end of the last epoch.
-        if epoch % self.gif_and_model_save_frequency != 0:
-            # Generate a GIF of all the saved images
-            self.generate_gif()
+        self.save_model_checkpoint()
+        # epoch = len(self.metrics_dataframe)
+        # # skip generating a gif if training ends on a multiple of the frequency to create a gif because it was already generated in the 
+        # # generate_validation_samples method at the end of the last epoch.
+        # if epoch % self.gif_and_model_save_frequency != 0:
+        #     # Generate a GIF of all the saved images
+        #     self.generate_gif()
         return
     
     def update_current_epoch(self):
